deduplicate main()
